:sparkle: created new Adaptors Functionality module :art: moved map providers to the 'providers' package :art: updated what's new
