+++ AdaptorsFunctionality/manifest.mf
@@ -1,0 +1,6 @@
+Manifest-Version: 1.0
+AutoUpdate-Show-In-Client: true
+OpenIDE-Module: au.gov.asd.tac.constellation.functionality.adaptors
+OpenIDE-Module-Localizing-Bundle: au/gov/asd/tac/constellation/functionality/adaptors/Bundle.properties
+OpenIDE-Module-Specification-Version: 1.0
+
