+++ AdaptorsFunctionality/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=febd8b19
+build.xml.script.CRC32=fd083cd6
+build.xml.stylesheet.CRC32=a56c6a5b@2.72.1
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=febd8b19
+nbproject/build-impl.xml.script.CRC32=56497b52
+nbproject/build-impl.xml.stylesheet.CRC32=238281d1@2.72.1
