+++ AdaptorsFunctionality/nbproject/project.properties
