+++ AdaptorsFunctionality/nbproject/suite.properties
@@ -1,0 +1,1 @@
+suite.dir=${basedir}/..
