+++ AdaptorsFunctionality/src/au/gov/asd/tac/constellation/functionality/adaptors/Bundle.properties
@@ -1,0 +1,1 @@
+OpenIDE-Module-Name=Adaptors Functionality
