+++ AdaptorsMapView/nbproject/project.properties
@@ -1,2 +1,2 @@
-javac.source=1.7
+javac.source=1.8
 javac.compilerargs=-Xlint -Xlint:-serial
