+++ AdaptorsMapView/src/au/gov/asd/tac/constellation/views/mapview/adaptors/providers/GoogleMapProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package au.gov.asd.tac.constellation.views.mapview.adaptors;
+package au.gov.asd.tac.constellation.views.mapview.adaptors.providers;
 
 import au.gov.asd.tac.constellation.views.mapview.providers.MapProvider;
 import de.fhpotsdam.unfolding.core.Coordinate;
@@ -21,23 +21,38 @@ import org.openide.util.lookup.ServiceProvider;
 import processing.core.PImage;
 
 /**
- * OpenTopoMap map.
+ * Google Maps map.
  * 
  * @author cygnus_x-1
  */
-@ServiceProvider(service = MapProvider.class, position = Integer.MAX_VALUE - 3)
-public class OpenTopoMapProvider extends MapProvider {
+@ServiceProvider(service = MapProvider.class, position = Integer.MAX_VALUE - 4)
+public class GoogleMapProvider extends MapProvider {
+    
+    private static final String LAYER_ROADS_ONLY = "h";
+    private static final String LAYER_STREETMAP = "m";
+    private static final String LAYER_STREETMAP_TERRAIN = "p";
+    private static final String LAYER_STREETMAP_ALT = "r";
+    private static final String LAYER_SATELLITE_ONLY = "s";
+    private static final String LAYER_TERRAIN_ONLY = "t";
+    private static final String LAYER_HYBRID = "y";
     
     @Override
     public String getName() {
-        return "OpenTopoMap";
+        return "Google Maps";
     }
 
     @Override
     public int zoomLevels() {
-        return 17;
+        return 20;
     }
     
+    @Override
+    public String getZoomString(final Coordinate coordinate) {
+        return "x=" + (int) coordinate.column
+                + "&y=" + (int) coordinate.row
+                + "&z=" + (int) coordinate.zoom;
+    }
+
     @Override
     public PImage getTile(Coordinate coordinate) {
         return null;
@@ -45,10 +60,10 @@ public class OpenTopoMapProvider extends MapProvider {
 
     @Override
     public String[] getTileUrls(Coordinate coordinate) {
-        // TODO: supply a special user agent string or this will be blocked
+        // TODO: supply an api key or this will be blocked
         final String url = String.format(
-                "https://tile.opentopomap.org/%s.png", 
-                getZoomString(coordinate));
+                "http://mt.google.com/vt/lyrs=%s&%s", 
+                LAYER_STREETMAP, getZoomString(coordinate));
         return new String[]{url};
     }
 }
