+++ AdaptorsMapView/src/au/gov/asd/tac/constellation/views/mapview/adaptors/providers/OpenTopoMapProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package au.gov.asd.tac.constellation.views.mapview.adaptors;
+package au.gov.asd.tac.constellation.views.mapview.adaptors.providers;
 
 import au.gov.asd.tac.constellation.views.mapview.providers.MapProvider;
 import de.fhpotsdam.unfolding.core.Coordinate;
@@ -21,26 +21,21 @@ import org.openide.util.lookup.ServiceProvider;
 import processing.core.PImage;
 
 /**
- * Stamen map.
- *
+ * OpenTopoMap map.
+ * 
  * @author cygnus_x-1
  */
-@ServiceProvider(service = MapProvider.class, position = Integer.MAX_VALUE - 1)
-public class StamanMapProvider extends MapProvider {
+@ServiceProvider(service = MapProvider.class, position = Integer.MAX_VALUE - 3)
+public class OpenTopoMapProvider extends MapProvider {
     
-    private static final String LAYER_TONER = "toner";
-    private static final String LAYER_TONER_BACKGROUND = "toner-background";
-    private static final String LAYER_TONER_LITE = "toner-lite";
-    private static final String LAYER_WATERCOLOR = "watercolor";
-
     @Override
     public String getName() {
-        return "Stamen Maps";
+        return "OpenTopoMap";
     }
 
     @Override
     public int zoomLevels() {
-        return 16;
+        return 17;
     }
     
     @Override
@@ -50,9 +45,10 @@ public class StamanMapProvider extends MapProvider {
 
     @Override
     public String[] getTileUrls(Coordinate coordinate) {
+        // TODO: supply a special user agent string or this will be blocked
         final String url = String.format(
-                "http://tile.stamen.com/%s/%s.png",
-                LAYER_TONER, getZoomString(coordinate));
+                "https://tile.opentopomap.org/%s.png", 
+                getZoomString(coordinate));
         return new String[]{url};
     }
 }
