+++ nbproject/project.properties
@@ -7,8 +7,12 @@ auxiliary.org-netbeans-modules-apisupport-installer.os-solaris=false
 auxiliary.org-netbeans-modules-apisupport-installer.os-windows=true
 auxiliary.org-netbeans-modules-apisupport-installer.pack200-enabled=true
 auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
-modules=${project.au.gov.asd.tac.constellation.views.mapview.adaptors}\:${project.au.gov.asd.tac.constellation.dependencies.adaptors}
+modules=\
+    ${project.au.gov.asd.tac.constellation.views.mapview.adaptors}:\
+    ${project.au.gov.asd.tac.constellation.dependencies.adaptors}:\
+    ${project.au.gov.asd.tac.constellation.functionality.adaptors}
 project.au.gov.asd.tac.constellation.dependencies.adaptors=AdaptorsDependencies
+project.au.gov.asd.tac.constellation.functionality.adaptors=AdaptorsFunctionality
 project.au.gov.asd.tac.constellation.views.mapview.adaptors=AdaptorsMapView
 
 ivy.classpath=AdaptorsDependencies/release/modules/ext/unfolding-master-20171010.jar
